communication/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor>=3.0
python-dotenv>=1.0.0